Minor changes in working directories
Changes happened during check of absence data issue in GBIF data. Re-run of the data preparation lead to same number of occurrence records (occ_points); thus, no influence on analysis is expected. Issue was about the occurrenceStatus column in GBIF data, that was not filtered for Present only but included Absent records.